Tests/LibWeb/Text/expected/Streams/ReadableStream-pipeTo.txt

abcdefghijklmnopqrstuvwxyz
ABCDEFGHIJKLMNOPQRSTUVWXYZ
0123456789!@#$%^&*()-=_+,<
